Split api/annotations
