--- src/components/consts.ts
@@ -1,11 +1,0 @@
-export enum AnnotationPriorities {
-  NORMAL = 'NORMAL',
-  WARNING = 'WARNING',
-  ALERT = 'ALERT',
-}
-
-export const annotationPrioritiesLabels = {
-  NORMAL: 'dodatkowa informacja',
-  WARNING: 'doprecyzowanie',
-  ALERT: 'sprostowanie błędu',
-};
