+++ src/store/reducer.ts
@@ -3,7 +3,7 @@ import { reducer as api } from 'redux-json-api';
 import widgets, { WidgetReducer } from './widgets/reducers';
 import appModes, { AppModeReducer } from './appModes/reducers';
 import textSelector from './textSelector/reducers';
-import { AnnotationAPIModel, AnnotationUpvoteAPIModel } from 'api/annotations';
+import { AnnotationAPIModel, AnnotationUpvoteAPIModel } from 'api';
 
 export interface ITabState {
   api: {
