+++ src/store/widgets/reducers.ts
@@ -7,7 +7,7 @@ import {
 } from './actions';
 import _difference from 'lodash/difference';
 import { API_DELETED } from 'redux-json-api/lib/constants';
-import { AnnotationResourceType } from '../../api/annotations';
+import { AnnotationResourceType } from 'api';
 import { combineReducers } from 'redux';
 
 export interface IWidgetState {
